add fluent api to validation request
