+++ AISAM.API/Validators/LoginUserDtoValidator.cs
@@ -1,0 +1,19 @@
+using BookStore.Common.Models;
+using FluentValidation;
+
+namespace BookStore.API.Validators
+{
+    public class LoginUserDtoValidator : AbstractValidator<LoginUserDto>
+    {
+        public LoginUserDtoValidator()
+        {
+            RuleFor(x => x.EmailOrUsername)
+                .NotEmpty().WithMessage("Email hoặc Username là bắt buộc");
+
+            RuleFor(x => x.Password)
+                .NotEmpty().WithMessage("Mật khẩu là bắt buộc");
+        }
+    }
+}
+
+
