+++ AISAM.Common/Models/AuthDtos.cs
@@ -1,0 +1,49 @@
+namespace BookStore.Common.Models
+{
+    public class AuthUrlResponse
+    {
+        public string AuthUrl { get; set; } = string.Empty;
+        public string State { get; set; } = string.Empty;
+    }
+
+    public class RegisterUserDto
+    {
+        public string Email { get; set; } = string.Empty;
+        public string Username { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
+        public string ConfirmPassword { get; set; } = string.Empty;
+    }
+
+    public class LoginUserDto
+    {
+        public string EmailOrUsername { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
+    }
+
+    public class UserDto
+    {
+        public int Id { get; set; }
+        public string? Email { get; set; }
+        public string? Username { get; set; }
+        public DateTime CreatedAt { get; set; }
+        public List<SocialAccountDto> SocialAccounts { get; set; } = new();
+    }
+
+    public class UserResponseDto
+    {
+        public int Id { get; set; }
+        public string Email { get; set; } = string.Empty;
+        public string Username { get; set; } = string.Empty;
+        public DateTime CreatedAt { get; set; }
+        public List<SocialAccountDto> SocialAccounts { get; set; } = new();
+    }
+
+    public class LoginResponseDto
+    {
+        public string Token { get; set; } = string.Empty;
+        public DateTime ExpiresAt { get; set; }
+        public UserResponseDto User { get; set; } = new();
+    }
+}
+
+
